+++ app/scripts/util/locale.js
@@ -101,6 +101,12 @@ var Locale = {
     openSelectFile: 'Select a file',
     openSelectFileBody: 'Select a file from your Dropbox which you would like to open',
     openPassFor: 'Password for',
+    openRemoveLastQuestion: 'Delete local file?',
+    openRemoveLastQuestionBody: 'The file you are deleting is stored only in browser. Delete it permanently?',
+    openLocalFile: 'Local file',
+    openLocalFileBody: 'You are going to open a local file. Changes you make will not be saved back to file system. ' +
+        'To get the file with your data, export it from settings.',
+    openLocalFileDontShow: 'Don\'t show this again',
 
     detAttDownload: 'Shift-click attachment button to download or ',
     detAttDelToRemove: 'Delete to remove',